move healthcheck out of admin
