+++ hawc/main/settings/base.py
@@ -39,6 +39,9 @@ MANAGERS = ADMINS
 ADMIN_URL_PREFIX = os.getenv("ADMIN_URL_PREFIX", "f09ea0b8-c3d5-4ff9-86c4-27f00e8f643d")
 ADMIN_ROOT = os.environ.get("ADMIN_ROOT", "")
 
+# add randomness to healthcheck url
+HEALTHCHECK_URL_PREFIX = os.getenv("HEALTHCHECK_URL_PREFIX", "4df02970-c4f8-4e16-8a75-e9eaf4dabac7")
+
 # {PRIME, EPA}
 HAWC_FLAVOR = os.getenv("HAWC_FLAVOR", "PRIME")
